requirements.txt
spacy==3.7.4
nltk==3.8.1
scikit-learn==1.4.1.post1
PyPDF2==3.0.1
python-docx==1.1.0
